backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djoser==2.1.0
django-filter==2.3.0
Pillow==9.0.0
short-url==1.2.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
flake8==6.0.0
flake8-isort==6.0.0